MULE-1697 - More testing before starting the refactoring.  This works quite well - can transfer 1GB without problems while running in IntelliJ IDEA limited to several 100MB memory (just takes time...).
git-svn-id: file:///home/projects/mule/scm@6324 bf997673-6b11-0410-b953-e057580c5b09

+++ trunk/mule/tests/functional/src/main/java/org/mule/tck/functional/FunctionalStreamingTestComponent.java
@@ -43,14 +43,14 @@ public class FunctionalStreamingTestComponent implements StreamingService
     public static final int STREAM_BUFFER_SIZE = 4096;
     private EventCallback eventCallback;
     private String summary = null;
-    private int targetSize = -1;
+    private long targetSize = -1;
 
     public FunctionalStreamingTestComponent()
     {
         logger.debug("creating " + toString());
     }
 
-    public void setEventCallback(EventCallback eventCallback, int targetSize)
+    public void setEventCallback(EventCallback eventCallback, long targetSize)
     {
         logger.debug("setting callback: " + eventCallback + " in " + toString());
         this.eventCallback = eventCallback;
@@ -72,7 +72,7 @@ public class FunctionalStreamingTestComponent implements StreamingService
             byte[] endData = new byte[STREAM_SAMPLE_SIZE]; // ring buffer
             int endDataSize = 0;
             int endRingPointer = 0;
-            int streamLength = 0;
+            long streamLength = 0;
             byte[] buffer = new byte[STREAM_BUFFER_SIZE];
 
             // throw data on the floor, but keep a record of size, start and end values
@@ -121,7 +121,7 @@ public class FunctionalStreamingTestComponent implements StreamingService
 
     private void doCallback(byte[] startData, int startDataSize,
                             byte[] endData, int endDataSize, int endRingPointer,
-                            int streamLength, UMOEventContext context) throws Exception
+                            long streamLength, UMOEventContext context) throws Exception
     {
         // make a nice summary of the data
         StringBuffer result = new StringBuffer("Received stream");
